requirements.txt
requests~=2.31.0
selenium~=4.14.0
openai~=0.28.1
python-dotenv~=1.0.1